Refactor server forms and blueprints + WIP plex configs
