+++ client/src/contexts/AuthContext.js
@@ -121,8 +121,9 @@ const AuthContextProvider = (props) => {
     }
   };
 
-  const signInWithPlex = async () => {
-    const res = await executeRequest(methods.GET, "/sign-in/plex/");
+  const signInWithPlex = async (redirectURI = null) => {
+    const uri = redirectURI ? redirectURI : "";
+    const res = await executeRequest(methods.GET, "/sign-in/plex/?redirectURI=" + uri);
     switch (res.status) {
       case 200:
         window.location.href = res.headers.location;
@@ -137,6 +138,7 @@ const AuthContextProvider = (props) => {
     const res = await executeRequest(methods.GET, "/plex/authorize/" + search);
     switch (res.status) {
       case 200:
+        console.log(res);
         initSession(res.data.username, res.data["user_picture"]);
         return res;
       default:
@@ -375,7 +377,19 @@ const AuthContextProvider = (props) => {
         return res;
       default:
         handleError(res);
+        return null;
+    }
+  };
+
+  const getPlexConfig = async () => {
+    const res = await executeRequest(methods.GET, "/provider/plex/config/");
+    switch (res.status) {
+      case 200:
+      case 404:
         return res;
+      default:
+        handleError(res);
+        return null;
     }
   };
 
@@ -405,6 +419,7 @@ const AuthContextProvider = (props) => {
       resetApiKey,
       deleteApiKey,
       deleteAccount,
+      getPlexConfig
     }}>
       { isLoadingSession && <PageLoader/> }
       { props.children }
