+++ requirements.txt
@@ -2,16 +2,18 @@ blinker
 cloudinary
 Flask
 Flask-Cors
+Flask-WTF
 Flask-Limiter
 Flask-Login
 Flask-Marshmallow
 Flask-Migrate
 Flask-SQLAlchemy
 Flask-Talisman
-Flask-WTF
 gunicorn
 itsdangerous
 marshmallow-sqlalchemy
+passlib
+PlexAPI
 psycopg2-binary
 pytest
 pytest-mock
@@ -19,3 +21,5 @@ requests
 sendgrid
 sqlalchemy
 Werkzeug
+wtforms
+WTForms-Alchemy
