+++ server/profile/routes/user_profile.py
@@ -6,15 +6,15 @@ from flask_login import fresh_login_required, current_user, login_required
 from server import InvalidUsage, db, utils, limiter
 from server.auth.models import User
 from server.auth.forms import PasswordForm, EmailForm
-from server.profile import profile
+from server.profile.routes import profile
 from server.profile.forms import UsernameForm, ChangePasswordForm, PictureForm
-from server.profile.serializers.user_serializer import user_serializer
+from server.profile.serializers.profile_serializer import profile_serializer
 
 
 @profile.route("/", methods=["GET"])
 @login_required
 def get_profile():
-    return user_serializer.dump(current_user), HTTPStatus.OK
+    return profile_serializer.dump(current_user), HTTPStatus.OK
 
 
 @profile.route("/", methods=["DELETE"])
@@ -28,7 +28,7 @@ def delete_user():
             status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
         )
 
-    if not current_user.check_password(password_form.password.data):
+    if not current_user.password == password_form.password.data:
         raise InvalidUsage("Wrong password.", HTTPStatus.BAD_REQUEST)
 
     current_user.delete()
@@ -64,11 +64,10 @@ def change_password():
             payload=password_form.errors,
         )
 
-    if not current_user.check_password(password_form.oldPassword.data):
+    if not current_user.password == password_form.oldPassword.data:
         raise InvalidUsage(
             "The passwords don't match.", status_code=HTTPStatus.BAD_REQUEST,
         )
-
     current_user.change_password(password_form.newPassword.data)
     html = render_template("email/change_password_notice.html")
     subject = "Your password has been changed"
